requirements.txt
absl-py==2.3.1
altair==5.5.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
catboost==1.2.8
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
colorclass==2.2.2
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
docopt==0.6.2
executing==2.2.0
flatbuffers==25.2.10
fonttools==4.59.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kagglehub==0.3.12
keras==3.11.2
kiwisolver==1.4.9
libclang==18.1.1
lightgbm==2.2.3
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
narwhals==2.1.2
nest-asyncio==1.6.0
numpy==2.3.2
opt_einsum==3.4.0
optree==0.17.0
packaging==24.2
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
pip-upgrader==1.4.15
platformdirs==4.3.8
plotly==6.3.0
prompt_toolkit==3.0.51
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.45.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminaltables==3.1.10
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
xgboost==0.90
